+++ test/unit-test.js
@@ -7,9 +7,16 @@ var ColorShortener = require('../lib/colors/shortener');
 var lineBreak = process.platform == 'win32' ? '\r\n' : '\n';
 var cssContext = function(groups, options) {
   var context = {};
-  var clean = function(cleanedCSS) {
+  var clean = function(expectedCSS) {
     return function(css) {
-      assert.equal(cleanCSS.process(css, options), cleanedCSS);
+      var cleanedCSS = null;
+      try {
+        cleanedCSS = cleanCSS.process(css, options);
+      } catch (e) {
+        // swallow - cleanedCSS is set to null and that's the new expected value
+      }
+
+      assert.equal(cleanedCSS, expectedCSS);
     };
   };
 
@@ -810,11 +817,11 @@ title']",
   '@import': cssContext({
     'empty': [
       "@import url();",
-      ""
+      null
     ],
     'of an unknown file': [
       "@import url('fake.css');",
-      ""
+      null
     ],
     'of a http file': "@import url(http://pro.goalsmashers.com/test.css);",
     'of a https file': [
@@ -829,7 +836,7 @@ title']",
     'of a remote file via // url with media': "@import url(//pro.goalsmashers.com/test.css) screen,tv;",
     'of a directory': [
       "@import url(test/data/partials);",
-      ""
+      null
     ],
     'of a real file': [
       "@import url(test/data/partials/one.css);",
@@ -915,7 +922,7 @@ title']",
   '@import with absolute paths': cssContext({
     'of an unknown file': [
       "@import url(/fake.css);",
-      ""
+      null
     ],
     'of a real file': [
       "@import url(/partials/one.css);",